requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.5
whitenoise==6.2.0

celery
coverage

django-crispy-forms
django-storages
django-tinymce



pre-commit
psycopg2-binary
pytest
pytest-cov
pytest-django
pytest-playwright
pytest-xdist
redis
